requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==6.0.0
    # via streamlit
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via click
cryptography==46.0.4
    # via google-auth
deprecation==2.1.0
    # via streamlit-modal
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-api-core==2.29.0
    # via google-api-python-client
google-api-python-client==2.188.0
    # via -r requirements.in
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.4
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via google-api-core
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via
    #   email-validator
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.15.0
    # via altair
numpy==2.4.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   altair
    #   deprecation
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.1.0
    # via streamlit
proto-plus==1.27.0
    # via google-api-core
protobuf==6.33.4
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via streamlit
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydeck==0.9.1
    # via streamlit
pyparsing==3.3.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.52.2
    # via
    #   -r requirements.in
    #   streamlit-modal
streamlit-modal==0.1.2
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   streamlit
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
validate-docbr==1.11.1
    # via -r requirements.in
watchdog==6.0.0
    # via streamlit
Faker